lab4_context_free_grammar/py_project/requirements.txt
numpy==1.22.4
Pillow==9.1.1